third_party/blink/web_tests/fast/dom/Document/createAttributeNS-namespace-err-expected.txt

createAttirbuteNS tests adapted from createElementNS tests attached to webkit bug 16833

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS createAttributeNS(undefined, undefined)
PASS createAttributeNS(null, undefined)
PASS createAttributeNS(undefined, null)
PASS createAttributeNS(null, null)
PASS createAttributeNS(null, ""); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided is empty.
PASS createAttributeNS("", null)
PASS createAttributeNS("", ""); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided is empty.
PASS createAttributeNS(null, "<div>"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided ('<div>') contains the invalid character '>'.
FAIL createAttributeNS(null, "0div")
PASS createAttributeNS(null, "di v"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided ('di v') contains the invalid character ' '.
FAIL createAttributeNS(null, "di<v")
FAIL createAttributeNS(null, "-div")
FAIL createAttributeNS(null, ".div")
PASS createAttributeNS("http://example.com/", "<div>"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided ('<div>') contains the invalid character '>'.
FAIL createAttributeNS("http://example.com/", "0div")
FAIL createAttributeNS("http://example.com/", "di<v")
FAIL createAttributeNS("http://example.com/", "-div")
FAIL createAttributeNS("http://example.com/", ".div")
FAIL createAttributeNS(null, ":div")
PASS createAttributeNS(null, "div:"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided ('div:') has an empty local name.
FAIL createAttributeNS("http://example.com/", ":div")
PASS createAttributeNS("http://example.com/", "div:"); threw InvalidCharacterError: Failed to execute 'createAttributeNS' on 'Document': The qualified name provided ('div:') has an empty local name.
PASS createAttributeNS(null, "d:iv"); threw NamespaceError: Failed to execute 'createAttributeNS' on 'Document': The namespace URI provided ('') is not valid for the qualified name provided ('d:iv').
FAIL createAttributeNS(null, "a:b:c"); valid XML name, invalid QName; expected INVALID_CHARACTER_ERR, threw NAMESPACE_ERR
FAIL createAttributeNS("http://example.com/", "a:b:c"); valid XML name, invalid QName
FAIL createAttributeNS(null, "a::c"); valid XML name, invalid QName; expected INVALID_CHARACTER_ERR, threw NAMESPACE_ERR
FAIL createAttributeNS("http://example.com/", "a::c"); valid XML name, invalid QName
FAIL createAttributeNS("http://example.com/", "a:0"); valid XML name, not a valid QName
FAIL createAttributeNS("http://example.com/", "0:a"); 0 at start makes it not a valid XML name
PASS createAttributeNS("http://example.com/", "a:_")
FAIL createAttributeNS("http://example.com/", "a:ெ"); non-ASCII character after colon is CombiningChar, which is NCNameChar but not (Letter | "_") so invalid at start of NCName (but still a valid XML name, hence not INVALID_CHARACTER_ERR)
FAIL createAttributeNS("http://example.com/", "ெ:a"); non-ASCII character after colon is CombiningChar, which is NCNameChar but not (Letter | "_") so invalid at start of NCName (Gecko chooses to throw NAMESPACE_ERR here, but either is valid as this is both an invalid XML name and an invalid QName)
PASS createAttributeNS("http://example.com/", "a:aெ")
PASS createAttributeNS("http://example.com/", "aெ:a")
PASS createAttributeNS("http://example.com/", "xml:test"); binding xml prefix wrong; threw NamespaceError: Failed to execute 'createAttributeNS' on 'Document': The namespace URI provided ('http://example.com/') is not valid for the qualified name provided ('xml:test').
PASS createAttributeNS("http://example.com/", "xmlns:test"); binding xmlns prefix wrong; threw NamespaceError: Failed to execute 'createAttributeNS' on 'Document': The namespace URI provided ('http://example.com/') is not valid for the qualified name provided ('xmlns:test').
PASS createAttributeNS("http://www.w3.org/2000/xmlns/", "x:test"); binding namespace namespace to wrong prefix; threw NamespaceError: Failed to execute 'createAttributeNS' on 'Document': The namespace URI provided ('http://www.w3.org/2000/xmlns/') is not valid for the qualified name provided ('x:test').
PASS createAttributeNS("http://www.w3.org/2000/xmlns/", "xmlns:test")
PASS createAttributeNS("http://www.w3.org/XML/1998/namespace", "xml:test")
PASS createAttributeNS("http://www.w3.org/XML/1998/namespace", "x:test")
PASS createAttributeNS("http://www.w3.org/2000/xmlns/", "xmlns")
PASS createAttributeNS("http://example.com/", "xmlns"); threw NamespaceError: Failed to execute 'createAttributeNS' on 'Document': The namespace URI provided ('http://example.com/') is not valid for the qualified name provided ('xmlns').
PASS successfullyParsed is true

TEST COMPLETE
